fix jdbc array handling, split handling for some array and multi value operator, split and add more tests
