+++ sql/src/main/java/org/apache/druid/sql/avatica/DruidStatement.java
@@ -28,6 +28,7 @@ import org.apache.calcite.avatica.Meta;
 import org.apache.calcite.avatica.remote.TypedValue;
 import org.apache.calcite.rel.type.RelDataType;
 import org.apache.calcite.rel.type.RelDataTypeField;
+import org.apache.calcite.sql.type.SqlTypeName;
 import org.apache.druid.java.util.common.ISE;
 import org.apache.druid.java.util.common.StringUtils;
 import org.apache.druid.java.util.common.concurrent.Execs;
@@ -114,12 +115,29 @@ public class DruidStatement implements Closeable
 
     for (int i = 0; i < fieldList.size(); i++) {
       RelDataTypeField field = fieldList.get(i);
-      final ColumnMetaData.Rep rep = QueryMaker.rep(field.getType().getSqlTypeName());
-      final ColumnMetaData.ScalarType columnType = ColumnMetaData.scalar(
-          field.getType().getSqlTypeName().getJdbcOrdinal(),
-          field.getType().getSqlTypeName().getName(),
-          rep
-      );
+
+      final ColumnMetaData.AvaticaType columnType;
+      if (field.getType().getSqlTypeName() == SqlTypeName.ARRAY) {
+        final ColumnMetaData.Rep elementRep = QueryMaker.rep(field.getType().getComponentType().getSqlTypeName());
+        final ColumnMetaData.ScalarType elementType = ColumnMetaData.scalar(
+            field.getType().getComponentType().getSqlTypeName().getJdbcOrdinal(),
+            field.getType().getComponentType().getSqlTypeName().getName(),
+            elementRep
+        );
+        final ColumnMetaData.Rep arrayRep = QueryMaker.rep(field.getType().getSqlTypeName());
+        columnType = ColumnMetaData.array(
+            elementType,
+            field.getType().getSqlTypeName().getName(),
+            arrayRep
+        );
+      } else {
+        final ColumnMetaData.Rep rep = QueryMaker.rep(field.getType().getSqlTypeName());
+        columnType = ColumnMetaData.scalar(
+            field.getType().getSqlTypeName().getJdbcOrdinal(),
+            field.getType().getSqlTypeName().getName(),
+            rep
+        );
+      }
       columns.add(
           new ColumnMetaData(
               i, // ordinal
