+++ sql/src/main/java/org/apache/druid/sql/calcite/expression/OperatorConversions.java
@@ -257,7 +257,7 @@ public class OperatorConversions
     /**
      * Sets the return type of the operator to "typeName", marked as non-nullable.
      *
-     * One of {@link #returnTypeNonNull}, {@link #returnTypeNullable}, or
+     * One of {@link #returnTypeNonNull}, {@link #returnTypeNullable}, {@link #returnTypeNullableArray}, or
      * {@link #returnTypeInference(SqlReturnTypeInference)} must be used before calling {@link #build()}. These methods
      * cannot be mixed; you must call exactly one.
      */
@@ -274,7 +274,7 @@ public class OperatorConversions
     /**
      * Sets the return type of the operator to "typeName", marked as nullable.
      *
-     * One of {@link #returnTypeNonNull}, {@link #returnTypeNullable}, or
+     * One of {@link #returnTypeNonNull}, {@link #returnTypeNullable}, {@link #returnTypeNullableArray}, or
      * {@link #returnTypeInference(SqlReturnTypeInference)} must be used before calling {@link #build()}. These methods
      * cannot be mixed; you must call exactly one.
      */
@@ -287,11 +287,28 @@ public class OperatorConversions
       );
       return this;
     }
+    /**
+     * Sets the return type of the operator to an array type with elements of "typeName", marked as nullable.
+     *
+     * One of {@link #returnTypeNonNull}, {@link #returnTypeNullable}, {@link #returnTypeNullableArray}, or
+     * {@link #returnTypeInference(SqlReturnTypeInference)} must be used before calling {@link #build()}. These methods
+     * cannot be mixed; you must call exactly one.
+     */
+    public OperatorBuilder returnTypeNullableArray(final SqlTypeName elementTypeName)
+    {
+      Preconditions.checkState(this.returnTypeInference == null, "Cannot set return type multiple times");
+
+      this.returnTypeInference = ReturnTypes.explicit(
+          factory -> Calcites.createSqlArrayTypeWithNullability(factory, elementTypeName, true)
+      );
+      return this;
+    }
+
 
     /**
      * Provides customized return type inference logic.
      *
-     * One of {@link #returnTypeNonNull}, {@link #returnTypeNullable}, or
+     * One of {@link #returnTypeNonNull}, {@link #returnTypeNullable}, {@link #returnTypeNullableArray}, or
      * {@link #returnTypeInference(SqlReturnTypeInference)} must be used before calling {@link #build()}. These methods
      * cannot be mixed; you must call exactly one.
      */
