+++ sql/src/main/java/org/apache/druid/sql/calcite/expression/builtin/ArrayAppendOperatorConversion.java
@@ -24,30 +24,33 @@ import org.apache.calcite.sql.SqlFunction;
 import org.apache.calcite.sql.SqlFunctionCategory;
 import org.apache.calcite.sql.SqlOperator;
 import org.apache.calcite.sql.type.OperandTypes;
+import org.apache.calcite.sql.type.ReturnTypes;
 import org.apache.calcite.sql.type.SqlTypeFamily;
-import org.apache.calcite.sql.type.SqlTypeName;
 import org.apache.druid.segment.column.RowSignature;
 import org.apache.druid.sql.calcite.expression.DruidExpression;
 import org.apache.druid.sql.calcite.expression.OperatorConversions;
 import org.apache.druid.sql.calcite.expression.SqlOperatorConversion;
 import org.apache.druid.sql.calcite.planner.PlannerContext;
 
-public class MultiValueStringPrependOperatorConversion implements SqlOperatorConversion
+public class ArrayAppendOperatorConversion implements SqlOperatorConversion
 {
   private static final SqlFunction SQL_FUNCTION = OperatorConversions
-      .operatorBuilder("MV_PREPEND")
+      .operatorBuilder("ARRAY_APPEND")
       .operandTypeChecker(
           OperandTypes.sequence(
-              "(expr,array)",
+              "(array,expr)",
               OperandTypes.or(
-                  OperandTypes.family(SqlTypeFamily.STRING),
-                  OperandTypes.family(SqlTypeFamily.NUMERIC)
+                  OperandTypes.family(SqlTypeFamily.ARRAY),
+                  OperandTypes.family(SqlTypeFamily.STRING)
               ),
-              OperandTypes.family(SqlTypeFamily.STRING)
+              OperandTypes.or(
+                OperandTypes.family(SqlTypeFamily.STRING),
+                OperandTypes.family(SqlTypeFamily.NUMERIC)
+              )
           )
       )
       .functionCategory(SqlFunctionCategory.STRING)
-      .returnTypeNonNull(SqlTypeName.VARCHAR)
+      .returnTypeInference(ReturnTypes.ARG0_NULLABLE)
       .build();
 
   @Override
@@ -69,7 +72,7 @@ public class MultiValueStringPrependOperatorConversion implements SqlOperatorCon
         rexNode,
         druidExpressions -> DruidExpression.of(
             null,
-            DruidExpression.functionCall("array_prepend", druidExpressions)
+            DruidExpression.functionCall("array_append", druidExpressions)
         )
     );
   }
