+++ sql/src/main/java/org/apache/druid/sql/calcite/expression/builtin/ArrayConcatOperatorConversion.java
@@ -24,28 +24,33 @@ import org.apache.calcite.sql.SqlFunction;
 import org.apache.calcite.sql.SqlFunctionCategory;
 import org.apache.calcite.sql.SqlOperator;
 import org.apache.calcite.sql.type.OperandTypes;
+import org.apache.calcite.sql.type.ReturnTypes;
 import org.apache.calcite.sql.type.SqlTypeFamily;
-import org.apache.calcite.sql.type.SqlTypeName;
 import org.apache.druid.segment.column.RowSignature;
 import org.apache.druid.sql.calcite.expression.DruidExpression;
 import org.apache.druid.sql.calcite.expression.OperatorConversions;
 import org.apache.druid.sql.calcite.expression.SqlOperatorConversion;
 import org.apache.druid.sql.calcite.planner.PlannerContext;
 
-public class StringToMultiValueStringOperatorConversion implements SqlOperatorConversion
+public class ArrayConcatOperatorConversion implements SqlOperatorConversion
 {
-  // note: since this function produces an array
   private static final SqlFunction SQL_FUNCTION = OperatorConversions
-      .operatorBuilder("STRING_TO_MV")
+      .operatorBuilder("ARRAY_CONCAT")
       .operandTypeChecker(
           OperandTypes.sequence(
-              "(string,expr)",
-              OperandTypes.family(SqlTypeFamily.STRING),
-              OperandTypes.family(SqlTypeFamily.STRING)
+              "(array,array)",
+              OperandTypes.or(
+                  OperandTypes.family(SqlTypeFamily.ARRAY),
+                  OperandTypes.family(SqlTypeFamily.STRING)
+              ),
+              OperandTypes.or(
+                  OperandTypes.family(SqlTypeFamily.ARRAY),
+                  OperandTypes.family(SqlTypeFamily.STRING)
+              )
           )
       )
       .functionCategory(SqlFunctionCategory.STRING)
-      .returnTypeNonNull(SqlTypeName.VARCHAR)
+      .returnTypeInference(ReturnTypes.ARG0_NULLABLE)
       .build();
 
   @Override
@@ -67,7 +72,7 @@ public class StringToMultiValueStringOperatorConversion implements SqlOperatorCo
         rexNode,
         druidExpressions -> DruidExpression.of(
             null,
-            DruidExpression.functionCall("string_to_array", druidExpressions)
+            DruidExpression.functionCall("array_concat", druidExpressions)
         )
     );
   }
