+++ sql/src/main/java/org/apache/druid/sql/calcite/expression/builtin/ArrayToStringOperatorConversion.java
@@ -40,9 +40,8 @@ public class ArrayToStringOperatorConversion implements SqlOperatorConversion
           OperandTypes.sequence(
               "(array,expr)",
               OperandTypes.or(
-                  OperandTypes.family(SqlTypeFamily.STRING),
                   OperandTypes.family(SqlTypeFamily.ARRAY),
-                  OperandTypes.family(SqlTypeFamily.MULTISET)
+                  OperandTypes.family(SqlTypeFamily.STRING)
               ),
               OperandTypes.family(SqlTypeFamily.ANY)
           )
