+++ sql/src/test/java/org/apache/druid/sql/avatica/DruidStatementTest.java
@@ -124,7 +124,7 @@ public class DruidStatementTest extends CalciteTestBase
     Assert.assertEquals(
         Lists.newArrayList(
             Lists.newArrayList("__time", "TIMESTAMP", "java.lang.Long"),
-            Lists.newArrayList("cnt", "BIGINT", "java.lang.Long"),
+            Lists.newArrayList("cnt", "BIGINT", "java.lang.Number"),
             Lists.newArrayList("dim1", "VARCHAR", "java.lang.String"),
             Lists.newArrayList("dim2", "VARCHAR", "java.lang.String"),
             Lists.newArrayList("dim3", "VARCHAR", "java.lang.String"),
